fix: double to int manager atts types

+++ src/main/java/acme/forms/ManagerDashboard.java
@@ -27,9 +27,9 @@ public class ManagerDashboard extends AbstractForm {
 
 	double						devUserStoryEstimatedCost;
 
-	double						minUserStoryEstimatedCost;
+	int							minUserStoryEstimatedCost;
 
-	double						maxUserStoryEstimatedCost;
+	int							maxUserStoryEstimatedCost;
 
 	double						avgProjectCost;
